+++ robotization dilemma/RenderOut.cpp
@@ -128,8 +128,8 @@ void drawXZPlane(float camcx, float camcy, float camcz, float d, int n) {
 
 void renderNewRound(int id) {
   reinterpret_cast<Graphics::LabelHwnd>(Graphics::getElementById("objectMainGameCommitButton"))->text = "Turn " + to_string(id);
-  reinterpret_cast<Graphics::SliderHwnd>(Graphics::getElementById("objectMainGameTimeSlider"))->maxv = ROUND_TIME*id;
-  reinterpret_cast<Graphics::SliderHwnd>(Graphics::getElementById("objectMainGameTimeSlider"))->val = ROUND_TIME*(id-1);
+  reinterpret_cast<Graphics::SliderHwnd>(Graphics::getElementById("objectMainGameTimeSlider"))->maxv = ROUND_TIME.toDouble()*id;
+  reinterpret_cast<Graphics::SliderHwnd>(Graphics::getElementById("objectMainGameTimeSlider"))->val = ROUND_TIME.toDouble()*(id-1);
 }
 
 void createMainMenu() {
